requirements.txt
twilio==6.45.0
python-dotenv==0.14.0
Flask==1.1.1
Flask-SocketIO==4.3.1
google==2.0.3
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
grpcio==1.31.0
numpy==1.17.2
